tikv_metrics_proxy: move ParseHostPortAddr func to pkg utils
