+++ tikv_metrics_proxy/main.go
@@ -1,15 +1,26 @@
+// Copyright 2018 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package main
 
 import (
 	"bytes"
 	"context"
 	"fmt"
-	"net"
 	"net/http"
 	"os"
 	"os/signal"
 	"sort"
-	"strings"
 	"sync"
 	"syscall"
 
@@ -32,25 +43,6 @@ var (
 
 type tikvOpts struct {
 	addrs string
-}
-
-// ParseHostPortAddr returns a host:port list
-func ParseHostPortAddr(s string) ([]string, error) {
-	strs := strings.Split(s, ",")
-	addrs := make([]string, 0, len(strs))
-
-	for _, str := range strs {
-		str = strings.TrimSpace(str)
-
-		_, _, err := net.SplitHostPort(str)
-		if err != nil {
-			return nil, errors.Annotatef(err, `tikv.addrs does not have the form "host:port": %s`, str)
-		}
-
-		addrs = append(addrs, str)
-	}
-
-	return addrs, nil
 }
 
 func checkFlags(opts tikvOpts) {
@@ -165,7 +157,7 @@ func main() {
 	log.Info("starting tikv_metrics_proxy")
 
 	var err error
-	stores, err = ParseHostPortAddr(opts.addrs)
+	stores, err = utils.ParseHostPortAddr(opts.addrs)
 	if err != nil {
 		log.Fatalf("initialize tikv_metrics_proxy error, %v", errors.ErrorStack(err))
 	}
